fix: 스케줄러로 변경
